requirements.txt
flask
flask-login
flask-openid
flask-mail
sqlalchemy
flask-sqlalchemy
flask-bootstrap
flask-migrate
sqlalchemy-migrate
flask-whooshalchemy
flask-wtf
pytz
flask-babel
flup
werkzeug
scikit-learn
pandas
numpy
pytest
pytest-cov
tensorflow
tensorflow-hub
pillow